small refactoring; small fixes
