+++ wine-management-api/src/main/java/org/example/winemanagementapi/dto/grape/GrapeRequest.java
@@ -1,4 +1,4 @@
-package org.example.winemanagementapi.dto;
+package org.example.winemanagementapi.dto.grape;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
