+++ wine-management-api/src/main/java/org/example/winemanagementapi/dto/grape/GrapeResponse.java
@@ -1,4 +1,4 @@
-package org.example.winemanagementapi.dto;
+package org.example.winemanagementapi.dto.grape;
 
 import lombok.Builder;
