+++ wine-management-api/src/main/java/org/example/winemanagementapi/dto/region/RegionResponse.java
@@ -1,4 +1,4 @@
-package org.example.winemanagementapi.dto;
+package org.example.winemanagementapi.dto.region;
 
 import lombok.Builder;
