+++ wine-management-api/src/main/java/org/example/winemanagementapi/mappers/GrapeMapper.java
@@ -1,16 +1,13 @@
 package org.example.winemanagementapi.mappers;
 
-import org.example.winemanagementapi.dto.GrapeRequest;
-import org.example.winemanagementapi.dto.GrapeResponse;
+import org.example.winemanagementapi.dto.grape.GrapeRequest;
+import org.example.winemanagementapi.dto.grape.GrapeResponse;
 import org.example.winemanagementapi.entities.Grape;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.mapstruct.factory.Mappers;
 
 @Mapper(componentModel = "spring")
 public interface GrapeMapper {
-
-    GrapeMapper INSTANCE = Mappers.getMapper(GrapeMapper.class);
 
     @Mapping(target = "id", ignore = true)
     @Mapping(target = "wines", ignore = true)
